# internal/api/api.yaml
openapi: 3.0.0
info:
  title: Payment Sandbox API
  version: 0.1.0
  description: Authentication APIs for merchant onboarding and login.
...
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://staging.api.example.com
    description: Staging server
  - url: https://api.example.com
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /merchant/refunds:
    post:
      operationId: RequestRefund
      summary: Merchant request refund
      tags: [Merchant]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestRefundRequest"
      responses:
        "201":
          description: Created
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: ListRefundHistory
      summary: Refund history
      tags: [Merchant]
      parameters:
        - in: query
          name: merchantId
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: pageSize
          schema: { type: integer }
      responses:
        "200":
          description: Paginated refund history
...
  /merchant/topups:
    post:
      operationId: RequestTopUp
      summary: Merchant request top-up
      tags: [Merchant]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestTopUpRequest"
      responses:
        "201":
          description: Created
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: ListTopUpHistory
      summary: Top-up history
      tags: [Merchant]
      parameters:
        - in: query
          name: merchantId
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: pageSize
          schema: { type: integer }
      responses:
        "200":
          description: Paginated top-up history
...
  /admin/refunds/{refundId}/approve:
    post:
      operationId: ApproveRefund
      summary: Admin approve refund
      tags: [Admin]
      parameters:
        - in: path
          name: refundId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Refund approved
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/refunds/{refundId}/reject:
    post:
      operationId: RejectRefund
      summary: Admin reject refund
      tags: [Admin]
      parameters:
        - in: path
          name: refundId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Refund rejected
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/refunds/{refundId}/process:
    post:
      operationId: ProcessRefund
      summary: Admin process refund outcome
      tags: [Admin]
      parameters:
        - in: path
          name: refundId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessRefundRequest"
      responses:
        "200":
          description: Refund processed
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RequestRefundRequest:
      type: object
      required: [invoiceId, merchantId, amount]
      properties:
        invoiceId:
          type: string
        merchantId:
          type: string
        amount:
          type: integer
          format: int64
    RequestTopUpRequest:
      type: object
      required: [merchantId, amount]
      properties:
        merchantId:
          type: string
        amount:
          type: integer
          format: int64
        requestKey:
          type: string
...
    ProcessRefundRequest:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Detailed error message